Now specify number of spikes in the star. The angle between each spike is calculated based on the count. ThickLine takes a float for angle in the constructor to be more flexible. Clean up SDL surfaces and filter table.

+++ Shapes2D/ThickLine.h
@@ -7,7 +7,7 @@
 class ThickLine : public Shape2D {
 
 public:
-   ThickLine(int x, int y, int length, int angle, int w, const Color& c1, const Color& c2);
+   ThickLine(int x, int y, int length, float angle, int w, const Color& c1, const Color& c2);
 
    virtual void draw(SDL_Surface* surf);
